# docs/swagger.yaml
basePath: /
definitions:
  domain.Role:
    enum:
    - admin
    - manager
    - employee
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleManager
    - RoleEmployee
  domain.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        description: Hashed
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      updatedAt:
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  http.CreateClientRequest:
    properties:
      document:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  http.CreateVehicleRequest:
    properties:
      brand:
        type: string
      client_id:
        type: string
      model:
        type: string
      plate:
        type: string
      year:
        type: integer
    type: object
...
  http.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  http.loginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
  http.refreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  http.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@autorepair.com
    name: API Support
  description: API for AutoRepair Shop Management System
  license:
    name: Proprietary
    url: http://autorepair.com/license
  termsOfService: http://swagger.io/terms/
  version: "1.0"
paths:
  /admin/clients:
    get:
      consumes:
      - application/json
      description: List all clients
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List Clients
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: Register a new client
      parameters:
      - description: Client Details
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/http.CreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create Client
      tags:
      - clients
...
  /admin/vehicles:
    get:
      consumes:
      - application/json
      description: List vehicles by client
      parameters:
      - description: Client ID
        in: query
        name: client_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Invalid client ID
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List Vehicles
      tags:
      - vehicles
    post:
      consumes:
      - application/json
      description: Register a new vehicle
      parameters:
      - description: Vehicle Details
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/http.CreateVehicleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create Vehicle
      tags:
      - vehicles
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get a JWT token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.loginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Use a refresh token to get a new access token
      parameters:
      - description: Refresh Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.refreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.loginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Refresh Access Token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, password and role
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"